+++ source/Core/Castor3D/Gui/Controls/CtrlLayoutControl.cpp
@@ -22,7 +22,7 @@ namespace castor3d
 		, ControlRPtr parent
 		, castor::Position const & position
 		, castor::Size const & size
-		, uint32_t flags
+		, uint64_t flags
 		, bool visible )
 		: Control{ type
 			, scene
